# backend/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: API do sistema de ressarcimento (SURE)
  title: SURE API
  version: "1.0"
paths:
...
  /api/v1/processos/{id}/emails:
    get:
      parameters:
      - description: ID do processo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Lista e-mails do processo
      tags:
      - EmailProcesso
    post:
      consumes:
      - application/json
      parameters:
      - description: ID do processo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Envia e-mail do processo
      tags:
      - EmailProcesso
  /api/v1/processos/{id}/emails/{emailId}/read:
    post:
      parameters:
      - description: ID do processo
        in: path
        name: id
        required: true
        type: integer
      - description: ID do e-mail
        in: path
        name: emailId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Marca e-mail do processo como lido (usuário atual)
      tags:
      - EmailProcesso
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"